# frontend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "互动教学平台 API",
    "description": "Interactive Teaching Platform Backend API Specification",
    "version": "1.0.0",
    "contact": {
      "name": "Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "开发环境"
    },
    {
      "url": "https://api.yourapp.com/api",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/discussions": {
      "get": {
        "tags": ["Discussions"],
        "summary": "获取讨论列表",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "courseId",
            "in": "query",
            "description": "课程ID筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "讨论类型",
            "schema": {
              "type": "string",
              "enum": ["public", "question"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取讨论列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DiscussionData"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Discussions"],
        "summary": "创建讨论帖",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscussionPostInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "讨论帖创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DiscussionPost"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/discussions/{postId}/like": {
      "put": {
        "tags": ["Discussions"],
        "summary": "点赞讨论帖",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "点赞成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "likes": {
                          "type": "integer"
                        },
                        "likedBy": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/discussions/{postId}/replies": {
      "post": {
        "tags": ["Discussions"],
        "summary": "添加回复",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "回复添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Reply"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "使用JWT token进行身份验证"
      }
    },
    "schemas": {
...
      "DiscussionData": {
        "type": "object",
        "properties": {
          "publicDiscussions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscussionPost"
            }
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscussionPost"
            }
          }
        }
      },
      "DiscussionPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "authorId": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "userRole": {
            "type": "string",
            "enum": ["teacher", "student", "ta"]
          },
          "isAnonymous": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": ["public", "question"]
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "likes": {
            "type": "integer"
          },
          "likedBy": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reply"
            }
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DiscussionPostInput": {
        "type": "object",
        "required": ["type", "title", "content"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["public", "question"]
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "isAnonymous": {
            "type": "boolean"
          }
        }
      },
      "Reply": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "authorId": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "userRole": {
            "type": "string",
            "enum": ["teacher", "student", "ta"]
          },
          "content": {
            "type": "string"
          },
          "likes": {
            "type": "integer"
          },
          "likedBy": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
...
  }
}